Create ui module and move MainActivity to presentation module
