+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,11 @@
         android:theme="@style/Theme.StarWarsApp"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.presentation.MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/Theme.StarWarsApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
