+++ presentation/src/main/kotlin/com/example/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.starwarsapp
+package com.example.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.starwarsapp.ui.theme.StarWarsAppTheme
+import com.example.ui.theme.StarWarsAppTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
