+++ settings.gradle.kts
@@ -24,3 +24,4 @@ include(":app")
 include(":data")
 include(":domain")
 include(":presentation")
+include(":ui")
