+++ ui/src/main/kotlin/com/example/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.starwarsapp.ui.theme
+package com.example.ui.theme
 
 import androidx.compose.ui.graphics.Color
