+++ ui/src/main/kotlin/com/example/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.example.starwarsapp.ui.theme
+package com.example.ui.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
