+++ ui/src/main/kotlin/com/example/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.starwarsapp.ui.theme
+package com.example.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
